requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
pyarrow==17.0.0
xgboost==2.1.1
category-encoders==2.6.3
joblib==1.4.2
requests==2.32.3
pyyaml==6.0.2